Added Knockback
Edited BonePhysics.cs

The player's velocity is now affected by projectiles.
Bones will now knock the player back.
If not grounded, the player's velocity is simply offset in the direction of the bone's velocity.
If grounded, the player "bounces" off the ground in the direction of the bone's velocity if it were to reflect off the ground.